Code improvements - proper file closing, file path checks, delete original optionality.

+++ zipping_instrument_raw_data.py
@@ -6,30 +6,36 @@ import time
 
 
 def get_directory():
-    directory = input('Please enter directory path E.g. data/home/instruments/raw_data:')
-    return directory
+    while True:
+        directory = input('Please enter directory path E.g. data/home/instruments/raw_data:')
+        if os.path.isdir(directory):
+            return directory
+        else:
+            print("The path entered is not a valid directory. Please try again.")
 
 
 def get_zip_targets(directory_name):
     target_directories = []
     for root, dirs, files in os.walk(directory_name, topdown=False):
         for dir in dirs:
-            target_directories.append(os.path.join(root, dir))
+            if not dir.endswith('.zip'):  # Skip directories ending with .zip
+                target_directories.append(os.path.join(root, dir))
     return target_directories
 
 
-def zip_the_targets(target_directories):
+def zip_the_targets(target_directories, delete_original=False):
     count = 0
     start_time = time.time()  # Store the start time
     for directory in target_directories:
         count += 1
-        zipf = zipfile.ZipFile(f'{directory}.zip', 'w', zipfile.ZIP_DEFLATED)
-        for root, dirs, files in os.walk(directory):
-            for file in files:
-                zipf.write(os.path.join(root, file), os.path.relpath(os.path.join(root, file), directory))
+        zip_path = f'{directory}.zip'
+        with zipfile.ZipFile(zip_path, 'w', zipfile.ZIP_DEFLATED) as zipf:
+            for root, dirs, files in os.walk(directory):
+                for file in files:
+                    zipf.write(os.path.join(root, file), os.path.relpath(os.path.join(root, file), directory))
 
-        zipf.close()
-        shutil.rmtree(directory)  # Delete original directory
+        if delete_original:
+            shutil.rmtree(directory)  # Delete original directory only if delete_original is True
         elapsed_time = time.time() - start_time  # Calculate elapsed time
         print(f'Directory {count} of {len(target_directories)} zipped. Elapsed time: {elapsed_time:.2f} seconds.')
 
@@ -38,7 +44,7 @@ def main():
     try:
         directory = get_directory()
         target_directories = get_zip_targets(directory)
-        zip_the_targets(target_directories)
+        zip_the_targets(target_directories, True)
         print('Zipping complete.')
 
     except Exception as e:
